missed one

+++ api/src/main/resources/sql/schemas/sc/sc.history.sql
@@ -13,7 +13,7 @@ create table sc.languages_ex_history (
 	province varchar(32),
 
 	first_language_population int,
-	population_value int, -- calculated from first_language_population
+	population_value decimal, -- calculated from first_language_population
 
 	egids_level sc.egids_scale,
 	egids_value decimal, -- calculated from egids_level
